# controllers/Event.js
'use strict';

var utils = require('../utils/writer.js');
var Event = require('../service/EventService');

module.exports.creatEvent = function creatEvent (req, res, next, body) {
  Event.creatEvent(body)
    .then(function (response) {
      utils.writeJson(res, response);
    })
    .catch(function (error) {
      // Αν υπάρχει σφάλμα, επιστρέφουμε το status 400 (Bad Request) ή το κατάλληλο status code
      res.status(400).json({ message: error.message });
    });
};

module.exports.view_a_specific_event = function view_a_specific_event(req, res, next, event_id) {
  Event.view_a_specific_event(event_id)
    .then(function (response) {
      utils.writeJson(res, response, 200);
    })
    .catch(function (error) {
      utils.writeJson(res, { message: error.message }, 404);
    });
};

module.exports.view_events = function view_events (req, res, next) {
  Event.view_events()
    .then(function (response) {
      utils.writeJson(res, response);
    })
    .catch(function (response) {
      utils.writeJson(res, response);
    });
};